refactor: sendFriendRequest adapter에서 viewModel로 이동
