+++ app/src/main/java/com/example/everymoment/presentation/adapter/FriendRequestAdapter.kt
@@ -1,6 +1,5 @@
 package com.example.everymoment.presentation.adapter
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.View.GONE
@@ -11,21 +10,15 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.everymoment.services.location.GlobalApplication
 import com.example.everymoment.R
-import com.example.everymoment.data.model.network.api.NetworkModule
-import com.example.everymoment.data.model.network.api.PotatoCakeApiService
 import com.example.everymoment.data.model.network.dto.response.Member
-import com.example.everymoment.data.model.network.dto.response.MemberResponse
 import com.example.everymoment.databinding.FriendRequestItemBinding
 import com.example.everymoment.extensions.CustomDialog
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import com.example.everymoment.presentation.viewModel.FriendRequestViewModel
 
 class FriendRequestAdapter(
     private val activity: FragmentActivity,
-    private val onFriendRequest: (Member) -> Unit
+    private val viewModel: FriendRequestViewModel
 ) : ListAdapter<Member, FriendRequestAdapter.FriendRequestViewHolder>(
     FriendRequestDiffCallback()
 ) {
@@ -33,7 +26,7 @@ class FriendRequestAdapter(
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FriendRequestViewHolder {
         val binding =
             FriendRequestItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return FriendRequestViewHolder(binding, activity, onFriendRequest)
+        return FriendRequestViewHolder(binding, activity, viewModel)
     }
 
     override fun onBindViewHolder(holder: FriendRequestViewHolder, position: Int) {
@@ -43,29 +36,33 @@ class FriendRequestAdapter(
     class FriendRequestViewHolder(
         private val binding: FriendRequestItemBinding,
         private val activity: FragmentActivity,
-        private val onFriendRequest: (Member) -> Unit
+        private val viewModel: FriendRequestViewModel
     ) : RecyclerView.ViewHolder(binding.root) {
-        private val apiService: PotatoCakeApiService = NetworkModule.provideApiService(NetworkModule.provideRetrofit())
-        private val jwtToken = GlobalApplication.prefs.getString("token", "null")
-        private val token = "Bearer $jwtToken"
         fun bind(user: Member) {
             binding.userNickname.text = user.nickname
 
             if (user.friendRequestStatus == "FRIEND") {
                 binding.friendButton.visibility = VISIBLE
                 binding.friendRequestButton.visibility = GONE
-            } else if(user.friendRequestStatus == "SELF") {
+                binding.requestCompletedButton.visibility = GONE
+            } else if (user.friendRequestStatus == "SELF") {
+                binding.requestCompletedButton.visibility = GONE
+                binding.friendButton.visibility = GONE
+                binding.friendRequestButton.visibility = GONE
+            } else if (user.friendRequestStatus == "SENT") {
+                binding.requestCompletedButton.visibility = VISIBLE
                 binding.friendButton.visibility = GONE
                 binding.friendRequestButton.visibility = GONE
             } else {
                 binding.friendButton.visibility = GONE
                 binding.friendRequestButton.visibility = VISIBLE
+                binding.requestCompletedButton.visibility = GONE
             }
 
             if (user.profileImageUrl == null) {
                 binding.profile.setImageResource(R.drawable.account_circle_24px)
             } else {
-                binding.profile.setPadding(15,15,15,15)
+                binding.profile.setPadding(15, 15, 15, 15)
                 Glide.with(itemView.context)
                     .load(user.profileImageUrl)
                     .circleCrop()
@@ -78,44 +75,12 @@ class FriendRequestAdapter(
         }
 
 
-
         private fun showFriendRequestConfirmationDialog(user: Member) {
             CustomDialog("${user.nickname}님에게\n친구 신청을 하시겠습니까?", "취소", "신청", onPositiveClick = {
-                sendFriendRequest(user.id) { success, _ ->
-                    if (success) {
-                        activity.runOnUiThread {
-                            onFriendRequest(user)
-                            binding.friendRequestButton.visibility = View.GONE
-                            binding.requestCompletedButton.visibility = View.VISIBLE
-                        }
-                    } else {
-                        activity.runOnUiThread {
-
-                        }
-                    }
-                }
+                viewModel.sendFriendRequest(user.id)
+                binding.friendRequestButton.visibility = GONE
+                binding.requestCompletedButton.visibility = VISIBLE
             }).show(activity.supportFragmentManager, "CustomDialog")
-        }
-
-        fun sendFriendRequest(
-            memberId: Int,
-            callback: (Boolean, MemberResponse?) -> Unit
-        ) {
-            apiService.sendFriendRequest(token, memberId).enqueue(object : Callback<MemberResponse> {
-                override fun onResponse(p0: Call<MemberResponse>, p1: Response<MemberResponse>) {
-                    if (p1.isSuccessful) {
-                        Log.d("FriendRequestPost", "${p1.body()}")
-                        callback(true, p1.body())
-                    } else {
-                        callback(false, null)
-                    }
-                }
-
-                override fun onFailure(p0: Call<MemberResponse>, p1: Throwable) {
-                    Log.d("FriendRequestPost", "Failed to fetch diaries: ${p1.message}")
-                    callback(false, null)
-                }
-            })
         }
     }
